+++ index.js
@@ -89,8 +89,6 @@ module.exports = {
     parse: function(text, opts) {
       return parser.parse(text, opts);
     },
-    cleanAST: parser.cleanAST,
-    massageAST: parser.massageAST,
     formatAST: function(ast, opts) {
       opts = normalizeOptions(opts);
       const doc = printAstToDoc(ast, opts);
